+++ Assembly-CSharp/Global/BGSCENE_DEF.cs
@@ -221,8 +221,8 @@ public class BGSCENE_DEF
         }
         else if (operation == "Position" && arguments.Length >= 3)
         {
-            Int16.TryParse(arguments[0], out bgOverlay.orgX);
-            Int16.TryParse(arguments[1], out bgOverlay.orgY);
+            float.TryParse(arguments[0], out bgOverlay.orgX);
+            float.TryParse(arguments[1], out bgOverlay.orgY);
             UInt16.TryParse(arguments[2], out bgOverlay.orgZ);
             bgOverlay.curX = bgOverlay.orgX;
             bgOverlay.curY = bgOverlay.orgY;
@@ -1425,8 +1425,8 @@ public class BGSCENE_DEF
 
                 // okay guise let's get a tile
 
-                int grabX = (overlayInfo.curX + spriteInfo.offX - tileSystem.MinX) / 16;
-                int grabY = (overlayInfo.curY + spriteInfo.offY - tileSystem.MinY) / 16;
+                int grabX = ((short)overlayInfo.curX + spriteInfo.offX - tileSystem.MinX) / 16;
+                int grabY = ((short)overlayInfo.curY + spriteInfo.offY - tileSystem.MinY) / 16;
                 Tile memoriaTile = memoriaOverlay.GetTile(grabX, grabY);
 
                 copyHelper.CopyTile(atlasArray, atlasSide, atlasX, atlasY, memoriaTile, memoriaOverlay, false);
